requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
flake8==5.0.4
Markdown==3.4.1
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
